requirements.txt
selenium==4.7.2
discord.py==2.1.0
python_dotenv==0.21.0
bs4==0.0.1
gspread==5.7.2
gspread_formatting==1.1.2
aiohttp==3.8.3
ebaysdk==2.2.0
pandas==1.5.2